+++ atlasdb-dbkvs-tests/src/test/java/com/palantir/atlasdb/keyvalue/dbkvs/DbkvsPostgresSerializableTransactionTest.java
@@ -15,28 +15,15 @@
  */
 package com.palantir.atlasdb.keyvalue.dbkvs;
 
-import org.junit.Ignore;
-
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
-import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionManagerAwareDbKvs;
-import com.palantir.atlasdb.keyvalue.impl.AbstractAtlasDbKeyValueServiceTest;
+import com.palantir.atlasdb.transaction.impl.AbstractSerializableTransactionTest;
 
-public class DbkvsKeyValueServiceTest extends AbstractAtlasDbKeyValueServiceTest {
+public class DbkvsPostgresSerializableTransactionTest extends
+        AbstractSerializableTransactionTest {
+
     @Override
     protected KeyValueService getKeyValueService() {
-        KeyValueService kvs = ConnectionManagerAwareDbKvs.create(DbkvsTestSuite.getKvsConfig());
-        for (TableReference table : kvs.getAllTableNames()) {
-            if (!table.getQualifiedName().equals("_metadata")) {
-                kvs.dropTable(table);
-            }
-        }
-        return kvs;
-    }
-
-    @Override
-    @Ignore
-    public void testGetRangeWithHistory() {
-        /* Have to ignore this test as it is an unsupported operation for this KVS */
+        return ConnectionManagerAwareDbKvs.create(DbkvsPostgresTestSuite.getKvsConfig());
     }
 }
